fix(download): relocate download pdf button
